+++ lithops/storage/backends/aws_s3/config.py
@@ -34,3 +34,8 @@ def load_config(config_data):
                 raise Exception(msg)
 
         config_data['aws_s3'].update(config_data['aws'])
+
+        if 'storage_bucket' not in config_data['aws_s3']:
+            key = config_data['aws_s3']['access_key_id']
+            region = config_data['aws_s3']['region_name']
+            config_data['aws_s3']['storage_bucket'] = f'lithops-bucket-{region}-{key[:6].lower()}'