+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [AWS EC2] Allow to automatically create the VPC and all the necessary resources
 - [IBM VPC & AWS EC2] General fixes and Improvements
 - [Executor] Allow to pass the config file location in the FunctionExecutor()
+- [Storage] Automatically create the storage bucket if not provided in the config
 
 ### Changed
 - [AWS EC2] Reduced number of mandatory parameters